requirements.txt
# Core dependencies
openai-whisper
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# Web downloads (opcional para MVP)
yt-dlp>=2024.1.0

# Utilities
python-dotenv==1.0.0
tqdm>=4.65.0
rich>=13.0.0  # Para CLI bonito

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.0
pre-commit==3.5.0

# Future API dependencies (comentadas por enquanto)
# fastapi==0.104.1
# uvicorn==0.24.0
# celery==5.3.4
# redis==5.0.1
# python-multipart==0.0.6